Switch to Serial Forwarding or WiFi with BOOT
Use BOOT button to switch from/to serial forwarding mode (motors can
only be driven through USB port) or to/from WiFi mode.

Default mode is set to Serial Forwarding.
Note that default Serial Forwarding baudrate is 921600 and not 1M like
for motors, because this is the maximum baudrate the CP210X USB-Serial
converter can handle.

Signed-off-by: Gilles DOFFE <g.doffe@gmail.com>